+++ src/activity/Activity.js
@@ -79,7 +79,6 @@ export default function Activity(Behaviour, activityDef, context) {
     getActivityById,
     getState,
     init,
-    message: inboundMessage,
     recover,
     resume,
     run,
@@ -680,10 +679,6 @@ export default function Activity(Behaviour, activityDef, context) {
       broker: broker.getState(),
       execution: execution && execution.getState(),
     };
-  }
-
-  function inboundMessage(messageContent) {
-    broker.publish('api', `activity.message.${messageContent.id}`, messageContent);
   }
 
   function next() {