+++ src/definition/DefinitionExecution.js
@@ -302,6 +302,7 @@ export default function DefinitionExecution(definition) {
       executionId,
       stopped,
       completed,
+      status,
       processes: processes.map((p) => p.getState()),
     };
   }
@@ -333,11 +334,11 @@ export default function DefinitionExecution(definition) {
     const content = message.content;
     const {target, source} = content;
 
-    logger.debug(`<${executionId} (${id})> conveying message from <${source.processId}.${source.id}> to <${target.processId}.${target.id}>`);
+    logger.debug(`<${executionId} (${id})> conveying message from <${source.processId}.${source.id}> to`, target.id ? `<${target.processId}.${target.id}>` : `<${target.processId}>`);
 
     const targetProcess = getProcessById(target.processId);
 
-    targetProcess.sendMessage(content);
+    targetProcess.sendMessage(message);
   }
 
   function onDelegateMessage(routingKey, executeMessage) {
@@ -381,7 +382,7 @@ export default function DefinitionExecution(definition) {
   }
 
   function getApi(apiMessage) {
-    if (!apiMessage) apiMessage = initMessage;
+    if (!apiMessage) apiMessage = initMessage || {content: createMessage()};
 
     const content = apiMessage.content;
     if (content.executionId !== executionId) {